requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
Pillow==10.0.1
python-magic==0.4.27
exifread==3.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.24.3
python-magic-bin